fix: 🐛 merge utils
